feat: Ajusdo envio do token

+++ app/services/api.ts
@@ -1,24 +1,31 @@
 import axios from 'axios';
 import { parseCookies, setCookie, destroyCookie } from 'nookies';
 
-const api = axios.create({
+const { 'erp.token': token } = parseCookies();
+
+const apiConfig = {
   baseURL: 'http://localhost:8082',
   timeout: 10000,
-});
+  headers: {},
+};
 
-let isRetryRefreshToken = false; 
+if (token) {
+  apiConfig.headers = {
+    Authorization: `Bearer ${token}`,
+  };
+}
+
+const api = axios.create(apiConfig);
+
+let isRetryRefreshToken = false;
 
 api.interceptors.response.use(
-  function (config) {
-    const { 'erp.token': token } = parseCookies();
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-    return config;
+  (response) => {
+    return response;
   },
   async function (error) {
 
-    if(isRetryRefreshToken) {
+    if (error.response?.status === 401 && isRetryRefreshToken) {
       destroyCookie(undefined, 'erp.token');
       destroyCookie(undefined, 'erp.refreshtoken');
       window.location.href = '/login';
@@ -27,22 +34,23 @@ api.interceptors.response.use(
     const { 'erp.token': access_token } = parseCookies();
     const originalRequest = error.config;
     if (error.response?.status === 401 && access_token) {
-        isRetryRefreshToken = true
-        const { 'erp.refreshtoken': refreshToken } = parseCookies();
-        const parameters = {
-          method: 'POST',
-        };
-        const body = {
-          refreshToken,
-        };
-        const response = await api.post('/auth/refresh/', body, parameters);
+      isRetryRefreshToken = true
+      const { 'erp.refreshtoken': refreshToken } = parseCookies();
+      const parameters = {
+        method: 'POST',
+      };
+      const body = {
+        refreshToken,
+      };
+      const response = await api.post('/auth/refresh/', body, parameters);
 
-        setCookie(undefined, 'erp.token', response.data.accessToken, {
-          maxAge: 60 * 60 * 1, // 1 hora
-        });
-        originalRequest.headers.Authorization = `Bearer ${response.data.accessToken}`;
-        return axios(originalRequest);
+      setCookie(undefined, 'erp.token', response.data.accessToken, {
+        maxAge: 60 * 60 * 1, // 1 hora
+      });
+      originalRequest.headers.Authorization = `Bearer ${response.data.accessToken}`;
+      return axios(originalRequest);
     }
+    isRetryRefreshToken = false;
     return Promise.reject(error);
   }
 );
